feat: 디자인 변경사항에 따라 구조 작성 및 기능 추가
