+++ src/page/common/applyMyInfoStep/FirstPage.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+import { StepProps } from '@/types/step.type';
+
+const FirstPage = ({ setIsFinishPage }: StepProps) => {
+  // const [nameStatus, setNameStatus] = useState<'success' | 'error' | 'default'>(
+  //   'default',
+  // );
+  // const [statusValue, setStatusValue] = useState('');
+  // const [nameValue, handleNameValue] = useInput('');
+  // // 초기화 필요 setIsFinishPage(false);
+  // const onClickDoubleCheckButton = () => {
+  //   setNameStatus('error');
+  //   setStatusValue('이미 있는 이름입니다.');
+  //   setIsFinishPage(true);
+  // };
+  return <></>;
+};
+
+export default FirstPage;
