+++ src/page/common/applyMyInfoStep/SecondPage.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { StepProps } from '@/types/step.type';
+
+const SecondPage = ({ setIsFinishPage }: StepProps) => {
+  return <></>;
+};
+
+export default SecondPage;
