+++ src/page/groupLeader/SecondStep/index.tsx
@@ -1,27 +1,27 @@
 'use client';
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
-import { Footer } from '@/components';
+import { Footer, ProgressHeader } from '@/components';
 
 import { useAppSelector } from '@/store/hooks';
 import { ApplyInfoState } from '@/store/feature/applyInfo';
 
 import FirstPage from './FirstPage';
 import SecondPage from './SecondPage';
-import ThirdPage from './ThirdPage';
-import FourthPage from './FourthPage';
-import FifthPage from './FifthPage';
-import SixthPage from './SixthPage';
 
-import { GROUP_MAX_PAGE_ARR } from '@/constants';
+import { GROUP_LEADER_MAX_PAGE_ARR } from '@/constants';
+import useCurPageReset from '@/hooks/useCurPageReset';
 
-const MAX_PAGE = GROUP_MAX_PAGE_ARR[1];
+const MAX_PAGE = GROUP_LEADER_MAX_PAGE_ARR[1];
 
-const GroupSecondStep = () => {
+const GroupLeaderSecondStep = () => {
   const { curPage } = useAppSelector(state => state.applyInfo);
 
   const [isFinishPage, setIsFinishPage] = useState(false);
+
+  // current Page 초기화
+  useCurPageReset({ curPage, setIsFinishPage });
 
   const changePage = (curPage: ApplyInfoState['curPage']) => {
     switch (curPage) {
@@ -29,15 +29,6 @@ const GroupSecondStep = () => {
         return <FirstPage setIsFinishPage={setIsFinishPage} />;
       case 2:
         return <SecondPage setIsFinishPage={setIsFinishPage} />;
-      case 3:
-        return <ThirdPage setIsFinishPage={setIsFinishPage} />;
-      case 4:
-        return <FourthPage setIsFinishPage={setIsFinishPage} />;
-      case 5:
-        return <FifthPage setIsFinishPage={setIsFinishPage} />;
-      case 6:
-        return <SixthPage setIsFinishPage={setIsFinishPage} />;
-
       default:
         return <></>;
     }
@@ -58,11 +49,11 @@ const GroupSecondStep = () => {
       {changePage(curPage)}
       <Footer
         maxPage={MAX_PAGE}
-        disabled={isFinishPage}
+        disabled={!isFinishPage}
         type={switchPageType()}
       />
     </>
   );
 };
 
-export default GroupSecondStep;
+export default GroupLeaderSecondStep;
