+++ src/page/groupLeader/ThirdStep/index.tsx
@@ -1,47 +1,68 @@
 'use client';
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import { Footer } from '@/components';
 
-import { useAppDispatch, useAppSelector } from '@/store/hooks';
+import { useAppSelector } from '@/store/hooks';
+import { ApplyInfoState } from '@/store/feature/applyInfo';
 
 import FirstPage from './FirstPage';
+import SecondPage from './SecondPage';
+import ThirdPage from './ThirdPage';
+import FourthPage from './FourthPage';
+import FifthPage from './FifthPage';
 
-import {
-  incrementStep,
-  decrementStep,
-  setPage,
-  resetPage,
-} from '@/store/feature/applyInfo';
+import { GROUP_LEADER_MAX_PAGE_ARR } from '@/constants';
+import useCurPageReset from '@/hooks/useCurPageReset';
 
-import { GROUP_MAX_PAGE_ARR } from '@/constants';
+const MAX_PAGE = GROUP_LEADER_MAX_PAGE_ARR[2];
 
-const MAX_PAGE = GROUP_MAX_PAGE_ARR[2];
+const GroupLeaderThirdStep = () => {
+  const { curPage } = useAppSelector(state => state.applyInfo);
 
-const GroupThirdStep = () => {
-  const { curStep } = useAppSelector(state => state.applyInfo);
-  const dispatch = useAppDispatch();
-  const onClickPrev = () => {
-    dispatch(setPage(GROUP_MAX_PAGE_ARR[curStep - 2]));
-    dispatch(decrementStep());
-  };
-  const onClickNext = () => {
-    dispatch(resetPage());
-    dispatch(incrementStep());
-  };
   const [isFinishPage, setIsFinishPage] = useState(false);
+
+  // current Page 초기화
+  useCurPageReset({ curPage, setIsFinishPage });
+
+  const changePage = (curPage: ApplyInfoState['curPage']) => {
+    switch (curPage) {
+      case 1:
+        return <FirstPage setIsFinishPage={setIsFinishPage} />;
+      case 2:
+        return <SecondPage setIsFinishPage={setIsFinishPage} />;
+      case 3:
+        return <ThirdPage setIsFinishPage={setIsFinishPage} />;
+      case 4:
+        return <FourthPage setIsFinishPage={setIsFinishPage} />;
+      case 5:
+        return <FifthPage setIsFinishPage={setIsFinishPage} />;
+      default:
+        return <></>;
+    }
+  };
+
+  const switchPageType = () => {
+    switch (curPage) {
+      case 1:
+        return 'firstPage';
+      case MAX_PAGE:
+        return 'lastPage';
+      default:
+        return;
+    }
+  };
   return (
     <>
-      <FirstPage setIsFinishPage={setIsFinishPage} />;
+      {changePage(curPage)}
       <Footer
         maxPage={MAX_PAGE}
-        disabled={isFinishPage}
-        onClickPrev={onClickPrev}
-        onClickNext={onClickNext}
+        disabled={!isFinishPage}
+        type={switchPageType()}
       />
     </>
   );
 };
 
-export default GroupThirdStep;
+export default GroupLeaderThirdStep;
